feat: show the inspectorlist on 'stations' tab
