--- packages/frontend/src/components/Modals/ReportsList/ReportsList.css
@@ -1,30 +1,0 @@
-.list-modal > div {
-    display: flex;
-    flex-direction: column;
-    margin-top: var(--margin-m);
-    border-bottom: var(--border-weak);
-}
-
-.list-modal > div:last-child {
-    border-bottom: none;
-}
-
-.list-modal {
-    height: 70vh;
-}
-
-.list-modal > div p {
-    color: var(--color-gray-on-dark-background);
-}
-
-.list-modal > div span {
-    color: var(--color-primary);
-}
-
-.list-modal > div > div {
-    justify-content: left;
-}
-
-.list-modal > div > div:first-child h4 {
-    margin-right: var(--margin-s);
-}
